+++ src/loggingprovider/loggingprovider.i.d.ts
