+++ src/samples/simple-freedom-chat/freedom-module.ts
@@ -2,16 +2,14 @@
 
 import Logging = require('../../logging/logging');
 
-import DataChannelInterfaces = require('../../webrtc/datachannel.i');
-import PeerConnectionInterfaces = require('../../webrtc/peerconnection.i');
+import WebRtcTypes = require('../../webrtc/types');
 import WebRtc = require('../../webrtc/webrtc');
 
-import PeerConnection = PeerConnectionInterfaces.PeerConnection;
-import SignallingMessage = PeerConnectionInterfaces.SignallingMessage;
-import PeerConnectionConfig = PeerConnectionInterfaces.PeerConnectionConfig;
-
-import DataChannel = DataChannelInterfaces.Channel;
-import Data = DataChannelInterfaces.Data;
+import PeerConnection = WebRtcTypes.PeerConnection;
+import SignallingMessage = WebRtcTypes.SignallingMessage;
+import PeerConnectionConfig = WebRtcTypes.PeerConnectionConfig;
+import DataChannel = WebRtcTypes.Channel;
+import Data = WebRtcTypes.Data;
 
 import Message = require('./messages.i');
 
@@ -96,7 +94,7 @@ a.negotiateConnection()
     parentFreedomModule.emit('ready', {});
     // Change logging tolerance once connected.
     // TODO: cleanup provider to show that we are supposed to do that.
-    parentFreedomModule['loggingprovider']().setConsoleFilter("*:I");
+    freedom['loggingcontroller']().setConsoleFilter("*:I");
   })
   .catch((e:any) => {
     log.error('error while opening datachannel: ' + e.message);
