+++ src/webrtc/datachannel.ts
@@ -10,12 +10,12 @@
 
 // import freedom = require('../../third_party/freedom/freedom-module-env');
 
-import DataChannels = require('./datachannel.i');
+import WebRtcTypes = require('./types');
 import Handler = require('../handler/queue');
 import ArrayBuffers = require('../arraybuffers/arraybuffers');
 import Logging = require('../logging/logging');
 
-import Data = DataChannels.Data;
+import Data = WebRtcTypes.Data;
 
 // Messages are limited to a 16KB length by SCTP; we use 15k for safety.
 // TODO: test if we can up this to 16k; test the edge-cases!
@@ -41,12 +41,12 @@ var log :Logging.Log = new Logging.Log('DataChannel');
 // Wrapper for a WebRtc Data Channels:
 // http://dev.w3.org/2011/webrtc/editor/webrtc.html#rtcdatachannel
 //
-class DataChannelClass implements DataChannels.Channel {
+class DataChannelClass implements WebRtcTypes.Channel {
 
   // Static, async constructor that returns a DataChannel ready for use.
   // |id| is the GUID generated by Freedom to identify the underlying
   //     browser object.
-  public static createFromFreedomId(id:string) : Promise<DataChannels.Channel> {
+  public static createFromFreedomId(id:string) : Promise<WebRtcTypes.Channel> {
     var rtcDataChannel = freedom['core.rtcdatachannel'](id);
     return rtcDataChannel.getLabel().then((label:string) => {
       return new DataChannelClass(rtcDataChannel, label, id);
